feat: reworked rt_timer logic to rung bechmark outside signal handler
Now benchmark execution happens as a standard function call and there is a semaphore wait which
determines when the next execution must happen. The next execution is triggered by a semaphore post
done by the new signal handler when the timer expires, if the previous execution has terminated.

BREAKING CHANGE: The signature of disparity_functions:dipsarity_execution is now the same as
disparity_init and disparity_teardown

Signed-off-by: Nick1296 <mattianicolella@gmail.com>