feat(periodic_benchmark): implement periods and logging functions
The general benchmark implementation (periodic_benchmark) has been reworked to handle periods. Now a
benchmark requires the definition of a period and a deadline, which must be less or equal to the
period. There is at exactly one deadline per period. Specifying a deadline or period can now be done
with a single option and using a decimal number (1,0.1,1.0,1.1e-8 are all examples of accepted
values). The reporting interface now computes also the benchmark elapsed time (in clock cycles), its
utilization, density and determines also if the job has met the deadline. Finally, the reporting
interface allows four level of verbosity and it influences how the reporting primitive works.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>