feat(base): add core affinity and output file name
With -c and the core id (or a comma-separated core id list) it is possible to specify on which core
the benchmark can run. In addition using -o the output file name must be specified. The default path
when -o is not used is ./timing.csv. Finally, log level 2 is the only log level that will write on
an output file.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>