Provide support for setting SCHED_FIFO and SCHED_DEADLINE policies
The new CLI parameters, -f, -P, -D, -T allow the user to setup a
FIFO/DEADLINE scheduling policy on Linux for each benchmark without
resorting of extra tools (e.g., chrt). The convention (for
SCHED_DEADLINE) about the parameters is the same as chrt (regarding the
implicit assignment of runtime, deadline, and period).

The setup uses sched_setattr syscall and passes down a in-benchmark
defined structure to allow extensions of the set of parameters on
custom kernels (e.g., for criticality).