fix(utils): normalize access to args parameter
the args object inside the params dictionary now is a dictionary. Additionally two new parameters
have been added to handle core pinning and it's now mandatory to set a priority

BREAKING CHANGE: Priority is mandatory and can be specifid either via the fifo or the deadline
parameters. The args key inside the param dictionary is now a dictionary and not a Namespace.