docs: Restrucutre of the documentation
Documentation is now up to date with the repo.
The repo itself has been restructured to follow more closely the
documentation.
The utility scripts have now proper modules and submodels that describe
them.
The CLI options for the benchmarks are described in the related modules
instead of being describe in the main page.

BREAKING CHANGE: The 'base' folder has been renamed to 'rt-bench_generator'.

Signed-off-by: Mattia Nicolella <mattianicolella@gmail.com>